remove simplecov_helper duplicate
